+++ librtl/src/lib_ichar.c
@@ -26,19 +26,18 @@
  * lib$ichar
  *
  */
-int lib$ichar(const struct dsc$descriptor_s* input_descriptor)
+unsigned int lib$ichar(const struct dsc$descriptor_s* input_descriptor)
 {
-    short word_integer_length;  /* Length of source string */
-    char* data_address;     /* Address of actual data */
-    long result = 0;        /* Result to return */
-    unsigned long status;
+    unsigned short word_integer_length; /* Length of source string */
+    char* data_address; /* Address of actual data */
+    int result = 0; /* Result to return */
+    int status;
 
     /*
      * use STR$ANALYSE_SDESC instead of LIB$ version, because
      * the LIB$ version calls the STR$ version.
      */
-    status = lib$analyze_sdesc(input_descriptor, &word_integer_length,
-                               &data_address);
+    status = lib$analyze_sdesc(input_descriptor, &word_integer_length, &data_address);
 
     if (word_integer_length >= 1)
     {
