+++ librtl/src/libmatchc.c
@@ -30,27 +30,27 @@
  */
 
 /*
-*
-*   Code for VAX LIB$MATCHC routine
-*
-* Description:
-*   Search a source string for a specified substring and returns an index
-*   which is the relative position of the first occurrence of a substring
-*   in the source string.
-*
-*   Returns 1,...,n for found string
-*       0 if not found
-*   If the substring has zero length return 1
-*   If the source string has zero length and the substring has a non-zero
-*   length - Zero is returned
-*
-* Bugs:
-*
-* History
-*
-*   Mar 25, 2005 - Andrew Allison
-*       Initial program creation - mainly copied from str$position
-*/
+ *
+ *   Code for VAX LIB$MATCHC routine
+ *
+ * Description:
+ *   Search a source string for a specified substring and returns an index
+ *   which is the relative position of the first occurrence of a substring
+ *   in the source string.
+ *
+ *   Returns 1,...,n for found string
+ *       0 if not found
+ *   If the substring has zero length return 1
+ *   If the source string has zero length and the substring has a non-zero
+ *   length - Zero is returned
+ *
+ * Bugs:
+ *
+ * History
+ *
+ *   Mar 25, 2005 - Andrew Allison
+ *       Initial program creation - mainly copied from str$position
+ */
 
 #include <stdio.h>
 #include <string.h>
@@ -59,14 +59,13 @@
 #include "str$routines.h"
 #include "lib$routines.h"
 
-unsigned long lib$matchc(const struct dsc$descriptor_s* search_string,
-                         const struct dsc$descriptor_s *source_string)
+unsigned int lib$matchc(const struct dsc$descriptor_s* search_string, const struct dsc$descriptor_s *source_string)
 {
-    char* s1_ptr;           /* Pointer to first string */
-    unsigned short s1_length;   /* Length of first string */
-    char* s2_ptr;           /* Pointer to second string */
-    unsigned short s2_length;   /* Length of second string */
-    long loop;          /* Outer loop */
+    char* s1_ptr; /* Pointer to first string */
+    unsigned short s1_length; /* Length of first string */
+    char* s2_ptr; /* Pointer to second string */
+    unsigned short s2_length; /* Length of second string */
+    unsigned int loop; /* Outer loop */
 
     /*
      * Analyze source strings
@@ -75,10 +74,10 @@ unsigned long lib$matchc(const struct dsc$descriptor_s* search_string,
     str$analyze_sdesc(source_string, &s2_length, &s2_ptr);
 
 //  Specical case search string is 0 and substring is non-zero
-    if (( s2_length == 0 ) && ( s1_length != 0 ))
+    if ((s2_length == 0) && (s1_length != 0))
         return 0;
 //  Special case if substring is NULL return 1
-    if ( s1_length == 0 )
+    if (s1_length == 0)
         return 1;
 
     /*
